removed everything except function relocation
