+++ doc/src/modules/matrices/expressions.rst
@@ -54,8 +54,6 @@ Matrix Expressions Core Reference
    :members:
 .. autoclass:: ZeroMatrix
    :members:
-.. autoclass:: Determinant
-   :members:
 
 Block Matrices
 --------------
