+++ doc/src/modules/matrices/index.rst
@@ -17,4 +17,3 @@ Contents:
    sparse.rst
    immutablematrices.rst
    expressions.rst
-   standalone.rst
