+++ sympy/matrices/__init__.py
@@ -24,15 +24,11 @@ SparseMatrix = MutableSparseMatrix
 from .expressions import (
     MatrixSlice, BlockDiagMatrix, BlockMatrix, FunctionMatrix, Identity,
     Inverse, MatAdd, MatMul, MatPow, MatrixExpr, MatrixSymbol, Trace,
-    Transpose, ZeroMatrix, OneMatrix, blockcut, block_collapse, matrix_symbols,
-    Adjoint, hadamard_product, HadamardProduct, HadamardPower, Determinant,
+    Transpose, ZeroMatrix, OneMatrix, blockcut, block_collapse, matrix_symbols, Adjoint,
+    hadamard_product, HadamardProduct, HadamardPower, Determinant, det,
     diagonalize_vector, DiagMatrix, DiagonalMatrix, DiagonalOf, trace,
     DotProduct, kronecker_product, KroneckerProduct,
     PermutationMatrix, MatrixPermute)
-
-from .determinant import (
-    adjugate, charpoly, cofactor, cofactor_matrix, det, det_bareiss,
-    det_berkowitz, det_LU, minor, minor_submatrix)
 
 __all__ = [
     'ShapeError', 'NonSquareMatrixError',
@@ -61,10 +57,7 @@ __all__ = [
     'MatrixSymbol', 'Trace', 'Transpose', 'ZeroMatrix', 'OneMatrix',
     'blockcut', 'block_collapse', 'matrix_symbols', 'Adjoint',
     'hadamard_product', 'HadamardProduct', 'HadamardPower', 'Determinant',
-    'diagonalize_vector', 'DiagMatrix', 'DiagonalMatrix',
+    'det', 'diagonalize_vector', 'DiagMatrix', 'DiagonalMatrix',
     'DiagonalOf', 'trace', 'DotProduct', 'kronecker_product',
     'KroneckerProduct', 'PermutationMatrix', 'MatrixPermute',
-
-    'adjugate', 'charpoly', 'cofactor', 'cofactor_matrix', 'det', 'det_bareiss',
-    'det_berkowitz', 'det_LU', 'minor', 'minor_submatrix',
 ]
