+++ sympy/matrices/matrices.py
@@ -84,6 +84,18 @@ class MatrixDeterminant(MatrixCommon):
         return _find_reasonable_pivot_naive(self, iszerofunc=iszerofunc,
                 simpfunc=simpfunc)
 
+    def _eval_det_bareiss(self, iszerofunc=_is_zero_after_expand_mul,
+            dotprodsimp=None):
+        return _det_bareiss(self, iszerofunc=iszerofunc,
+                dotprodsimp=dotprodsimp)
+
+    def _eval_det_berkowitz(self, dotprodsimp=None):
+        return _det_berkowitz(self, dotprodsimp=dotprodsimp)
+
+    def _eval_det_lu(self, iszerofunc=_iszero, simpfunc=None, dotprodsimp=None):
+        return _det_LU(self, iszerofunc=iszerofunc, simpfunc=simpfunc,
+                dotprodsimp=dotprodsimp)
+
     def _eval_determinant(self): # for expressions.determinant.Determinant
         return _det(self)
 
@@ -103,16 +115,6 @@ class MatrixDeterminant(MatrixCommon):
         return _det(self, method=method, iszerofunc=iszerofunc,
                 dotprodsimp=dotprodsimp)
 
-    def det_bareiss(self, iszerofunc=_is_zero_after_expand_mul, dotprodsimp=None):
-        return _det_bareiss(self, iszerofunc=iszerofunc, dotprodsimp=dotprodsimp)
-
-    def det_berkowitz(self, dotprodsimp=None):
-        return _det_berkowitz(self, dotprodsimp=dotprodsimp)
-
-    def det_LU(self, iszerofunc=_iszero, simpfunc=None, dotprodsimp=None):
-        return _det_LU(self, iszerofunc=iszerofunc, simpfunc=simpfunc,
-                dotprodsimp=dotprodsimp)
-
     def minor(self, i, j, method="berkowitz", dotprodsimp=None):
         return _minor(self, i, j, method=method, dotprodsimp=dotprodsimp)
 
@@ -121,15 +123,15 @@ class MatrixDeterminant(MatrixCommon):
 
     _find_reasonable_pivot.__doc__       = _find_reasonable_pivot.__doc__
     _find_reasonable_pivot_naive.__doc__ = _find_reasonable_pivot_naive.__doc__
+    _eval_det_bareiss.__doc__            = _det_bareiss.__doc__
+    _eval_det_berkowitz.__doc__          = _det_berkowitz.__doc__
+    _eval_det_lu.__doc__                 = _det_LU.__doc__
     _eval_determinant.__doc__            = _det.__doc__
     adjugate.__doc__                     = _adjugate.__doc__
     charpoly.__doc__                     = _charpoly.__doc__
     cofactor.__doc__                     = _cofactor.__doc__
     cofactor_matrix.__doc__              = _cofactor_matrix.__doc__
     det.__doc__                          = _det.__doc__
-    det_bareiss.__doc__                  = _det_bareiss.__doc__
-    det_berkowitz.__doc__                = _det_berkowitz.__doc__
-    det_LU.__doc__                       = _det_LU.__doc__
     minor.__doc__                        = _minor.__doc__
     minor_submatrix.__doc__              = _minor_submatrix.__doc__
